test-requirements.txt
importlib_metadata<5.0.0  # from kombu - https://github.com/celery/kombu/issues/1600
kombu
pyhamcrest
pytest
https://github.com/wazo-platform/wazo-test-helpers/archive/master.zip
